In the early stages of applying algorithms

+++ stat-bin/functions/average.go
@@ -2,16 +2,16 @@ package functions
 
 import "fmt"
 
-func Average(nms []int) int {
+func Average(nms []float64) float64 {
 	if len(nms) == 0 {
 		fmt.Println("No numbers")
 		return 0
 	}
 
-	sum := 0
+	sum := 0.0
 	for _, num := range nms {
 		sum += num
 	}
 
-	return sum / len(nms)
+	return sum / float64(len(nms))
 }
