+++ stat-bin/functions/median.go
@@ -2,14 +2,13 @@ package functions
 
 import "fmt"
 
-func Median(nms []int) int {
+func Median(nms []float64) float64 {
 	if len(nms) == 0 {
 		fmt.Println("No numbers to calculate median.")
 		return 0
 	}
 
 	res := Quicksort(nms)
-
 	mid := len(res) / 2
 	if len(res)%2 == 0 {
 		return (res[mid-1] + res[mid]) / 2
@@ -17,15 +16,15 @@ func Median(nms []int) int {
 	return res[mid]
 }
 
-func Quicksort(nms []int) []int {
+func Quicksort(nms []float64) []float64 {
 	if len(nms) <= 1 {
 		return nms
 	}
 
 	key := nms[0]
 
-	left := []int{}
-	right := []int{}
+	left := []float64{}
+	right := []float64{}
 
 	for i := 1; i < len(nms); i++ {
 		if key > nms[i] {
